refactor(shell/sidebar): extract bookmarks panel + dedupe getFaviconUrl
Move the BOOKMARKS PANEL MODULE (bmState, list/breadcrumb renderers,
add/edit/delete/folder flows, search) into shell/js/sidebar/panels/bookmarks.js,
exporting loadBookmarkPanel and BOOKMARK_PANEL_IDS. index.js drops from
1881 → ~1510 LOC.

Dedupe getFaviconUrl — PR #9 intentionally duplicated it between
history.js and index.js's bookmarks section with the expectation that
this PR would collapse it. New shared helper lives at
shell/js/sidebar/util.js; panels/bookmarks.js (new) and panels/history.js
(existing) both import from there.

No behavior change, no renames, no comment rewrites — pure extraction
per scope discipline §5.3, plus the one authorized dedupe.

Part of the 12-PR shell/js split (PR 10/12).
